+++ src/core/set.h
@@ -65,7 +65,7 @@ namespace Core
 
 		iterator insert(const KEY_TYPE& key)
 		{
-			const u32 keyHash = hasher_(0, key);
+			const u64 keyHash = hasher_(0, key);
 			const index_type indicesidx = keyHash & mask_;
 			const index_type idx = indices_[indicesidx];
 			if(idx != INVALID_INDEX)
@@ -118,7 +118,7 @@ namespace Core
 		{
 			if(indices_.size() > 0)
 			{
-				const u32 keyHash = hasher_(0, key);
+				const u64 keyHash = hasher_(0, key);
 				const index_type indicesidx = keyHash & mask_;
 				const index_type idx = indices_[indicesidx];
 				if(idx != INVALID_INDEX)
@@ -135,7 +135,7 @@ namespace Core
 		{
 			if(indices_.size() > 0)
 			{
-				const u32 keyHash = hasher_(0, key);
+				const u64 keyHash = hasher_(0, key);
 				const index_type indicesidx = keyHash & mask_;
 				const index_type idx = indices_[indicesidx];
 				if(idx != INVALID_INDEX)
@@ -169,7 +169,7 @@ namespace Core
 				// Reinsert all keys.
 				for(index_type idx = 0; idx < values_.size(); ++idx)
 				{
-					const u32 keyHash = hasher_(0, values_[idx]);
+					const u64 keyHash = hasher_(0, values_[idx]);
 					const index_type indicesidx = keyHash & mask_;
 					if(indices_[indicesidx] == INVALID_INDEX)
 					{
